ft-analyzer make use of statistic counters optional

+++ tools/ft-analyzer/ftanalyzer/models/statistical_model.py
@@ -126,6 +126,7 @@ class StatisticalModel:
         stats: GeneratorStats,
         log_dir: PathLike,
         merge: bool = False,
+        use_statistical_counter: bool = False,
         biflows_ts_correction: bool = False,
     ) -> None:
         """Read provided files and converts it to data frames.
@@ -214,24 +215,31 @@ class StatisticalModel:
         self._generator_stats: GeneratorStats = stats
         self._flows_ip_addresses_converted = False
         self._ref_ip_addresses_converted = isinstance(reference, pd.DataFrame)
+        self._stat_counter = use_statistical_counter
 
-        # statistic objects
-        self._sim = SimState(self._generator_stats.start_time)
+        if use_statistical_counter:
+            # statistic objects
+            self._sim = SimState(self._generator_stats.start_time)
 
-        output_dir = tempfile.mkdtemp()
+            output_dir = tempfile.mkdtemp()
 
-        self._statistic_objects, self._metric_to_obj = self._setup_statsitic_objects()
-        event_queue = create_event_queue(self._flows_path, output_dir)
-        self._process_events(event_queue, self._statistic_objects)
+            self._statistic_objects, self._metric_to_obj = (
+                self._setup_statsitic_objects()
+            )
+            event_queue = create_event_queue(self._flows_path, output_dir)
+            self._process_events(event_queue, self._statistic_objects)
 
-        shutil.rmtree(output_dir, ignore_errors=True)
-        output_dir = tempfile.mkdtemp()
+            shutil.rmtree(output_dir, ignore_errors=True)
+            output_dir = tempfile.mkdtemp()
 
-        self._ref_statisitic_objetcs, _ = self._setup_statsitic_objects()
-        ref_event_queue = create_event_queue(self._ref_path, output_dir)
-        self._process_events(ref_event_queue, self._ref_statisitic_objetcs)
+            self._ref_statisitic_objetcs, _ = self._setup_statsitic_objects()
+            ref_event_queue = create_event_queue(self._ref_path, output_dir)
+            self._process_events(ref_event_queue, self._ref_statisitic_objetcs)
 
-        shutil.rmtree(output_dir, ignore_errors=True)
+            shutil.rmtree(output_dir, ignore_errors=True)
+        else:
+            self._statistic_objects = {}
+            self._ref_statisitic_objetcs = {}
 
     def _load_flows_df(self):
         return pd.read_csv(
